4524 Always use verified downloads
